# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Warhammer Unit Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/kill-team/teams": {
      "get": {
        "tags": [
          "kill team"
        ],
        "summary": "List Kill Teams",
        "description": "Kill teams, alphabetically, optionally narrowed to one faction.\n\nAlphabetical because that is the only order there is: decision #25 gives print\norder to every CHILD of a kill team, not to the teams themselves, so the site's\nnav order is not stored.\n\nA `faction_id` matching nothing is an empty page, not a 404 -- a filter naming\nnothing is a legitimate result, where a missing ROW asked for by id is not.",
        "operationId": "list_kill_teams_api_v1_kill_team_teams_get",
        "parameters": [
          {
            "name": "faction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Faction Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_KillTeam_Read_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kill-team/teams/{kill_team_id}": {
      "get": {
        "tags": [
          "kill team"
        ],
        "summary": "Get Kill Team",
        "description": "One kill team with its rules, ploys, equipment, datacards and composition.\n\n404 for an unknown id, through the service's `NotFoundError` -- `app/main.py`\nregisters one handler against `CodedError`, so nothing is mapped here.",
        "operationId": "get_kill_team_api_v1_kill_team_teams__kill_team_id__get",
        "parameters": [
          {
            "name": "kill_team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Kill Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillTeam_Detail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/kill-team/rosters": {
      "post": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Create Roster",
        "description": "A new, empty roster. 409 if the player already has one by that name.",
        "operationId": "create_roster_api_v1_me_kill_team_rosters_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KTRoster_Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KTRoster_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "List Rosters",
        "description": "The caller's rosters, oldest first. Only ever the caller's.",
        "operationId": "list_rosters_api_v1_me_kill_team_rosters_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_KTRoster_ListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/kill-team/rosters/{roster_id}": {
      "get": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Get Roster",
        "operationId": "get_roster_api_v1_me_kill_team_rosters__roster_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KTRoster_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Update Roster",
        "operationId": "update_roster_api_v1_me_kill_team_rosters__roster_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KTRoster_Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KTRoster_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Delete Roster",
        "operationId": "delete_roster_api_v1_me_kill_team_rosters__roster_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/kill-team/rosters/{roster_id}/operatives": {
      "post": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Add Operative",
        "description": "Append one operative. A repeat is a second operative, not a 409.\n\nThe armies router's equivalent is create-only, because an incrementing add is not\nretry-safe. There is no quantity here (decision #50), so a retried add appends --\nwhich is a real outcome, and a caller that did not mean it removes the extra row.",
        "operationId": "add_operative_api_v1_me_kill_team_rosters__roster_id__operatives_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KTRosterOperative_Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KTRosterOperative_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/kill-team/rosters/{roster_id}/operatives/{row_id}": {
      "patch": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Move Operative",
        "description": "Set a row's position. Addressed by the ROW's id, not the operative's.",
        "operationId": "move_operative_api_v1_me_kill_team_rosters__roster_id__operatives__row_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Row Id"
            }
          },
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KTRosterOperative_Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KTRosterOperative_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kill team rosters"
        ],
        "summary": "Remove Operative",
        "operationId": "remove_operative_api_v1_me_kill_team_rosters__roster_id__operatives__row_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Row Id"
            }
          },
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Roster Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KTRosterOperative_Create": {
        "properties": {
          "operative_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operative Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operative_id"
        ],
        "title": "KTRosterOperative_Create"
      },
...
      "KTRoster_Create": {
        "properties": {
          "kill_team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Kill Team Id"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kill_team_id",
          "name"
        ],
        "title": "KTRoster_Create"
      },
...
      "KTRoster_Read": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "kill_team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Kill Team Id"
          },
          "kill_team_name": {
            "type": "string",
            "title": "Kill Team Name"
          },
          "faction_name": {
            "type": "string",
            "title": "Faction Name"
          },
          "operatives": {
            "items": {
              "$ref": "#/components/schemas/KTRosterOperative_Read"
            },
            "type": "array",
            "title": "Operatives",
            "default": []
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/KillTeamRule_Read"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          },
          "ploys": {
            "items": {
              "$ref": "#/components/schemas/KTPloy_Read"
            },
            "type": "array",
            "title": "Ploys",
            "default": []
          },
          "equipment": {
            "items": {
              "$ref": "#/components/schemas/KTEquipment_Read"
            },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "universal_ploys": {
            "items": {
              "$ref": "#/components/schemas/KTPloy_Read"
            },
            "type": "array",
            "title": "Universal Ploys",
            "default": []
          },
          "universal_equipment": {
            "items": {
              "$ref": "#/components/schemas/KTEquipment_Read"
            },
            "type": "array",
            "title": "Universal Equipment",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "kill_team_id",
          "kill_team_name",
          "faction_name"
        ],
        "title": "KTRoster_Read",
        "description": "A roster whole, as decision #52 specifies it.\n\nThe player's operatives with their datacards, then everything that applies while\nplaying them: the team's rules, its ploys, its equipment, and the two collections\nthat belong to no team. The equipment lists are the POOL a game chooses from \u2014 the\nchoice itself is per game (#17), so a roster holds no equipment of its own."
      },
      "KTRoster_Update": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "KTRoster_Update"
      },
...
      "KillTeam_Detail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "faction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faction Id"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/KillTeamRule_Read"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          },
          "ploys": {
            "items": {
              "$ref": "#/components/schemas/KTPloy_Read"
            },
            "type": "array",
            "title": "Ploys",
            "default": []
          },
          "equipment": {
            "items": {
              "$ref": "#/components/schemas/KTEquipment_Read"
            },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "operatives": {
            "items": {
              "$ref": "#/components/schemas/KTOperative_Read"
            },
            "type": "array",
            "title": "Operatives",
            "default": []
          },
          "selection_rules": {
            "items": {
              "$ref": "#/components/schemas/KTSelectionRule_Read"
            },
            "type": "array",
            "title": "Selection Rules",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "faction_id"
        ],
        "title": "KillTeam_Detail",
        "description": "A team whole: everything a reader of its page would see (decision #47).\n\nOne request rather than several, because operatives are reachable only nested and\na game snapshots the team's rules, its ploys and every datacard anyway (#24). Measured\nover the served responses: 17 KB to 47 KB, median 24 KB, each in nine queries flat."
      },
...
      "Page_KTRoster_ListRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KTRoster_ListRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[KTRoster_ListRead]"
      },
      "Page_KillTeam_Read_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KillTeam_Read"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[KillTeam_Read]"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}